Gather frames builder related code in a single module
- this makes the building of the frames buffer easier to follow
  and understand
  - merge in a single module code from: Frame.ts and TagsHelpers
- improve some names to something (hopefully) more appropriate
  - rename `createId3Data` to `createId3Tag`
  - rename `getSpecOptions` to `getFrameOptions`
  - rename `createBufferFromTags` to `buildFramesBuffer`
